Venta

+++ app/Http/Controllers/ProductoController.php
@@ -11,20 +11,21 @@ use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Storage;
 use App\Exports\ProductoExport;
-use PhpParser\Node\Stmt\TryCatch;
+
 
 class ProductoController extends Controller
 {
-    public function export()
+
+    public function export(Request $request)
     {
-        return Excel::download(new ProductoExport, 'productos.xlsx');
+        return Excel::download(new ProductoExport($request->columna, $request->orden), 'productos.xlsx');
     }
 
     public function index()
     {
         $datos['tipoProductos'] = TipoProducto::all();
 
-        $datos['imagenes']=ImagenProducto::all();
+        $datos['imagenes'] = ImagenProducto::all();
         $datos['Existencia'] = Producto::select("*")
             ->where("estado_producto", "=", 1)
             ->count();
@@ -53,27 +54,28 @@ class ProductoController extends Controller
         $input = request()->all();
 
         try {
+            
 
             DB::beginTransaction();
 
             $producto = Producto::insertGetId([
-                "nombre_producto" => $input['nombre_producto'],
-                "precio_producto" => $input['precio_producto'],
-                "estado_producto" => $input['estado_producto'],
-                "tipo_producto_id" => $input['tipo_producto_id'],
+                "nombre_producto" => $input['nombre_producto1'],
+                "precio_producto" => $input['precio_producto1'],
+                "estado_producto" => $input['estado_producto1'],
+                "tipo_producto_id" => $input['tipo_producto_id1'],
                 "existencia_producto" => 0
             ]);
 
-            if ($request->hasFile('url_imagen_producto')) {
+            if ($request->hasFile('url_imagen_producto1')) {
 
-                $imagenes = $request->file('url_imagen_producto');
+                $imagenes = $request->file('url_imagen_producto1');
                 foreach ($imagenes as $imagen) {
 
-                    
-                    
-                    $path=$imagen->store('uploads','public');
 
-                    
+
+                    $path = $imagen->store('uploads', 'public');
+
+
 
                     ImagenProducto::create([
                         'url_imagen_producto' => $path,
@@ -83,16 +85,17 @@ class ProductoController extends Controller
             }
 
             DB::commit();
-            return redirect("/producto")->with('status', '1');;
+            return redirect("/producto")->with('status', 'registrado');
         } catch (\Exception $e) {
             DB::rollBack();
 
-            $imagenes = $request->file('url_imagen_producto');
-            foreach ($imagenes as $imagen) {
+            if ($request->hasFile('url_imagen_producto1')) {
+                $imagenes = $request->file('url_imagen_producto1');
+                foreach ($imagenes as $imagen) {
 
-                Storage::delete('public/uploads'.$imagen->getClientOriginalName());
-                
-            }
+                    Storage::delete('public/uploads' . $imagen->getClientOriginalName());
+                }
+            } 
 
             return redirect("/producto")->with('status', $e->getMessage());
         }
@@ -100,7 +103,7 @@ class ProductoController extends Controller
 
     public function update(Request $request, $id)
     {
-        $this->validator($request->all())->validate();
+        $this->validatorUpdate($request->all())->validate();
         $input = request()->all();
 
         try {
@@ -108,17 +111,17 @@ class ProductoController extends Controller
 
 
             Producto::where('id_producto', '=', $id)->update([
-                "nombre_producto" => $input['nombre_producto'],
-                "precio_producto" => $input['precio_producto'],
-                "estado_producto" => $input['estado_producto'],
-                "tipo_producto_id" => $input['tipo_producto_id']
+                "nombre_producto" => $input['nombre_producto2'],
+                "precio_producto" => $input['precio_producto2'],
+                "estado_producto" => $input['estado_producto2'],
+                "tipo_producto_id" => $input['tipo_producto_id2']
             ]);
 
-            if ($request->hasFile('url_imagen_producto')) {
-                $imagenes = $request->file('url_imagen_producto');
+            if ($request->hasFile('url_imagen_producto2')) {
+                $imagenes = $request->file('url_imagen_producto2');
                 foreach ($imagenes as $imagen) {
-                   
-                    $path=$imagen->store('uploads','public');
+
+                    $path = $imagen->store('uploads', 'public');
                     ImagenProducto::create([
                         'url_imagen_producto' => $path,
                         'producto_id' =>  $id
@@ -127,16 +130,16 @@ class ProductoController extends Controller
             }
 
             DB::commit();
-            return redirect("/producto")->with('status', '1');
+            return redirect("/producto")->with('status', 'actualizado');
         } catch (\Exception $e) {
             DB::rollBack();
-            
-            $imagenes = $request->file('url_imagen_producto');
-            foreach ($imagenes as $imagen) {
+            if ($request->hasFile('url_imagen_producto2')) {
+                $imagenes = $request->file('url_imagen_producto2');
+                foreach ($imagenes as $imagen) {
 
-                Storage::delete('public/uploads'.$imagen->getClientOriginalName());
-                
-            }
+                    Storage::delete('public/uploads' . $imagen->getClientOriginalName());
+                }
+            } 
             return redirect("/producto")->with('status', $e->getMessage());
         }
     }
@@ -144,37 +147,40 @@ class ProductoController extends Controller
 
     public function destroy($id)
     {
-
         try {
             DB::beginTransaction();
             $prod = Producto::find($id);
+
             if ($prod->estado_producto == 0) {
                 Producto::where('id_producto', '=', $prod->id_producto)->update([
                     "estado_producto" => 1
                 ]);
+                $statusValidator = "restaurado";
             } else if ($prod->estado_producto == 1) {
                 Producto::where('id_producto', '=', $prod->id_producto)->update([
                     "estado_producto" => 0
                 ]);
+                $statusValidator = "cancelado";
             }
 
             DB::commit();
-            return redirect("/producto")->with('status', '1');
+            return redirect("/producto")->with('status', $statusValidator);
         } catch (\Exception $e) {
             DB::rollBack();
             return redirect("/producto")->with('status', $e->getMessage());
         }
     }
 
-    public function destroyImg($id){
-        $imagen=ImagenProducto::where('id_imagen_producto','=',$id)->first();
-        try{
+    public function destroyImg($id)
+    {
+        $imagen = ImagenProducto::where('id_imagen_producto', '=', $id)->first();
+        try {
             DB::beginTransaction();
-            ImagenProducto::where('id_imagen_producto','=',$id)->delete();
-            Storage::delete('public/'.$imagen->url_imagen_producto);
+            ImagenProducto::where('id_imagen_producto', '=', $id)->delete();
+            Storage::delete('public/' . $imagen->url_imagen_producto);
             DB::commit();
-            return redirect("/producto")->with('status', '1');
-        }catch(\Exception $e){
+            return redirect("/producto")->with('status', 'actualizado');
+        } catch (\Exception $e) {
             DB::rollBack();
             return redirect("/producto")->with('status', $e->getMessage());
         }
@@ -183,10 +189,23 @@ class ProductoController extends Controller
     public function validator(array $input)
     {
         return Validator::make($input, [
-            'nombre_producto' => ['required', 'max:26'],
-            'precio_producto' => ['required', 'int', 'max:999999', 'min:50'],
-            'url_imagen_producto.*' => ['image', 'mimes:jpeg,png,jpg,gif,svg', 'max:2048']
-
+            'nombre_producto1' => ['required', 'max:40'],
+            'precio_producto1' => ['required', 'int', 'max:999999', 'min:50'],
+            'url_imagen_producto1.*' => ['image', 'mimes:jpeg,png,jpg,gif,svg', 'max:2048'],
+            'tipo_producto_id1' => ['required','not_in:null'],
+            'estado_producto1'=>['required','not_in:null']
         ]);
     }
+    public function validatorUpdate(array $input)
+    {
+        return Validator::make($input, [
+           
+            'nombre_producto2' => ['required', 'max:40'],
+            'precio_producto2' => ['required', 'int', 'max:999999', 'min:50'],
+            'url_imagen_producto2.*' => ['image', 'mimes:jpeg,png,jpg,gif,svg', 'max:2048'],
+            'tipo_producto_id2' => ['required','not_in:null'],
+            'estado_producto2'=>['required','not_in:null']
+        ]);
+    }
+    
 }
